move the file collection function into a separate module
